+++ samples/shared/src/desktopMain/kotlin/com/github/panpf/sketch/sample/ui/test/DirectoriesTestViewModel.desktop.kt
@@ -8,42 +8,57 @@ import java.nio.file.Paths
 
 actual fun buildPlatformDirectoryItemList(context: PlatformContext): List<DirectoryItem> =
     buildList {
-        // macOS: Debug: hotReload: '/Users/panpf/Library/Caches'
-        // macOS: Debug: normal: '/Users/panpf/Library/Caches'
-        // macOS: Release: normal: '/Users/panpf/Library/Caches'
+        // windows: 'C:\Users\panpf\AppData\Local\Cache'
+        // macOS: '/Users/panpf/Library/Caches'
+        // linux: '/Users/panpf/.cache'
         val userCacheDirectory = AppDirsFactory.getInstance()
             .getUserCacheDir(/* appName = */ null,/* appVersion = */ null,/* appAuthor = */ null)
         add(DirectoryItem("userCacheDirectory", userCacheDirectory))
 
-        // macOS: Debug: hotReload: '/Users/panpf/Library/Application Support'
-        // macOS: Debug: normal: '/Users/panpf/Library/Application Support'
-        // macOS: Release: normal: '/Users/panpf/Library/Application Support'
+        // windows: 'C:\Users\panpf\AppData\Local'
+        // macOS: '/Users/panpf/Library/Application Support'
+        // linux: '/Users/panpf/.local/share'
         val userDataDirectory = AppDirsFactory.getInstance()
             .getUserDataDir(/* appName = */ null,/* appVersion = */ null,/* appAuthor = */ null)
         add(DirectoryItem("userDataDirectory", userDataDirectory))
 
-        // macOS: Debug: hotReload: '/Users/panpf/Library/Preferences'
-        // macOS: Debug: normal: '/Users/panpf/Library/Preferences'
-        // macOS: Release: normal: '/Users/panpf/Library/Preferences'
+        // windows: 'C:\Users\panpf\AppData\Roaming'
+        // macOS: '/Users/panpf/Library/Preferences'
+        // linux: '/Users/panpf/.config'
         val userConfigDirectory = AppDirsFactory.getInstance()
             .getUserConfigDir(/* appName = */ null,/* appVersion = */ null,/* appAuthor = */ null)
         add(DirectoryItem("userConfigDirectory", userConfigDirectory))
 
-        // macOS: Debug: hotReload: ''
-        // macOS: Debug: normal: '/Users/panpf/Workspace/sketch/samples/desktopApp/build/compose/tmp/prepareAppResources'
-        // macOS: Release: normal: '/Users/panpf/Downloads/Sketch4.app/Contents/app/resources'
+        // windows: Debug (normal): ''
+        // windows: Debug (hotReload): ''
+        // windows: Release: 'C:\Program Files\Sketch4\app\resources'
+        // macOS: Debug (normal): '/Users/panpf/Workspace/sketch/samples/desktopApp/build/compose/tmp/prepareAppResources'
+        // macOS: Debug (hotReload): ''
+        // macOS: Release: '/Users/panpf/Downloads/Sketch4.app/Contents/app/resources'
         val composeResourcesPath = getComposeResourcesPath()
         add(DirectoryItem("composeResourcesPath", composeResourcesPath))
 
-        // macOS: Debug. hotReload: '/Users/panpf/Workspace/sketch/samples/desktopApp/build/run/desktopMain/classpath/libs/sketch-core/9d92mnc/sketch-core-desktop-4.7.0-alpha01.jar'
-        // macOS: Debug. normal: '/Users/panpf/Workspace/sketch/sketch-core/build/libs/sketch-core-desktop-4.7.0-alpha01.jar'
-        // macOS: Release. normal: '/Users/panpf/Downloads/Sketch4.app/Contents/app/sketch-core-desktop-4.7.0-alpha01-6687c2e433cc6cca19d38f36afe9a9d4.jar'
+        // windows: Debug (normal): 'D:\Developer\Workspace\sketch\sketch-core\build\libs\sketch-core-desktop-4.7.0-alpha01.jar'
+        // windows: Debug (hotReload): 'D:\Developer\Workspace\sketch\samples\desktopApp\build\run\desktopMain\classpath\libs\sketch-core\9d92mnc\sketch-core-desktop-4.7.0-alpha01.jar'
+        // windows: Release: 'C:\Program Files\Sketch4\app\sketch-core-desktop-4.7.0-alpha01-6687c2e433cc6cca19d38f36afe9a9d4.jar'
+        // macOS: Debug (normal): '/Users/panpf/Workspace/sketch/sketch-core/build/libs/sketch-core-desktop-4.7.0-alpha01.jar'
+        // macOS: Debug (hotReload): '/Users/panpf/Workspace/sketch/samples/desktopApp/build/run/desktopMain/classpath/libs/sketch-core/9d92mnc/sketch-core-desktop-4.7.0-alpha01.jar'
+        // macOS: Release: '/Users/panpf/Downloads/Sketch4.app/Contents/app/sketch-core-desktop-4.7.0-alpha01-6687c2e433cc6cca19d38f36afe9a9d4.jar'
+        // linux: Debug (normal): '/home/panpf/Workspace/sketch/sketch-core/build/libs/sketch-core-desktop-4.7.0-alpha01.jar'
+        // linux: Debug (hotReload): '/home/panpf/Workspace/sketch/samples/desktopApp/build/run/desktopMain/classpath/libs/sketch-core/9d92mnc/sketch-core-desktop-4.7.0-alpha01.jar'
+        // linux: '/opt/hellokmp/lib/app/composeApp-desktop-e1e452276759301f909baa97e6a11ff4.jar'
         val jarPath = getJarPath(PlatformContext::class.java)
         add(DirectoryItem("jarPath", jarPath))
 
-        // macOS: Debug: hotReload: '/Users/panpf/Library/Caches/SketchImageLoader/ece624f7de942a10408f90b4ef73d207'
-        // macOS: Debug: normal: '/Users/panpf/Library/Caches/SketchImageLoader/a153ef8d1e6750f396f0582b25329ddb'
-        // macOS: Release: normal: '/Users/panpf/Library/Caches/SketchImageLoader/5fc3f0cb597bdebc54a2877d2c033920'
+        // windows: Debug (normal): 'C:\Users\panpf\AppData\Local\SketchImageLoader\a519cef9835e9be1cddb14770a1b6ff9\Cache'
+        // windows: Debug (hotReload): 'C:\Users\panpf\AppData\Local\SketchImageLoader\cb7785e11486b2233bb9f35f7bec2c77\Cache'
+        // windows: Release: 'C:\Users\panpf\AppData\Local\SketchImageLoader\e26c0c487c4bb4e45663c35c4400cae1\Cache'
+        // macOS: Debug (normal): '/Users/panpf/Library/Caches/SketchImageLoader/f99742e39ecdb5444589ad6aa2eff27a'
+        // macOS: Debug (hotReload): '/Users/panpf/Library/Caches/SketchImageLoader/3120c807db80a0e963f8c0073ec54f37'
+        // macOS: Release: '/Users/panpf/Library/Caches/SketchImageLoader/3a852cc448c41fefdf3b19c26497f70b'
+        // linux: Debug (normal): '/Users/panpf/.cache/SketchImageLoader/f99742e39ecdb5444589ad6aa2eff27a'
+        // linux: Debug (hotReload): '/Users/.cache/SketchImageLoader/3120c807db80a0e963f8c0073ec54f37'
+        // linux: Release: '/Users/panpf/.cache/SketchImageLoader/3a852cc448c41fefdf3b19c26497f70b'
         val appCacheDirectory = context.appCacheDirectory()?.toString()
         add(DirectoryItem("appCacheDirectory", appCacheDirectory))
     }